refactor makedDataObjectFrom
