+++ src/data/makeDataObjectFrom.js
